Create makeTagDirectory.slurm
This is a script used for making Tag directory. This process is required if you use Homer as the peak finder. The script first creates a directory for each chromosome. Then this script finds "accepted hits" while annotating which chromosome this hit comes from at the same time. The "accepted hit" means a mapped NGS read. Finally, it allocates each "accepted hits" to the chromosome directory it created before.